:hammer: Refactor Extraction code of StudyViewURL Query
